refactor(core): move game elements info directory
